pyproject.toml
[project]
name = "pytest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite==0.22.1",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.13.0",
    "asyncio==4.0.0",
    "bcrypt==4.3.0",
    "certifi==2026.2.25",
    "charset-normalizer==3.4.6",
    "chatollama==0.2.13",
    "click==8.3.1",
    "colorama==0.4.6",
    "fastapi==0.135.2",
    "greenlet==3.3.2",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "jsonpatch==1.33",
    "jsonpointer==3.1.1",
    "langchain==1.2.13",
    "langchain-core==1.2.23",
    "langchain-ollama==1.0.1",
    "langgraph==1.1.3",
    "langgraph-checkpoint==4.0.1",
    "langgraph-prebuilt==1.0.8",
    "langgraph-sdk==0.3.12",
    "langsmith==0.7.22",
    "ollama==0.6.1",
    "orjson==3.11.7",
    "ormsgpack==1.12.2",
    "packaging==26.0",
    "passlib>=1.7.4",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pyjwt==2.12.1",
    "pyyaml==6.0.3",
    "requests==2.33.0",
    "requests-toolbelt==1.0.0",
    "sqlalchemy==2.1.0b1",
    "starlette==1.0.0",
    "tenacity==9.1.4",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "uuid-utils==0.14.1",
    "uvicorn==0.42.0",
    "xxhash==3.6.0",
    "zstandard==0.25.0",
]
